Safari-611.1.10/LayoutTests/fast/url/segments-from-data-url2-expected.txt

Test URL segmentation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS segments('http://user:pass@foo:21/bar;par?b#c') is '["http:","foo","21","/bar;par","?b","#c"]'
PASS segments('http:foo.com') is '["http:","foo.com","","/","",""]'
PASS segments('\t   :foo.com   \n') is '[":","","","","",""]'
PASS segments(' foo.com  ') is '[":","","","","",""]'
PASS segments('a:\t foo.com') is '["a:","",""," foo.com","",""]'
PASS segments('http://f:21/ b ? d # e ') is '["http:","f","21","/%20b%20","?%20d%20","#%20e"]'
PASS segments('http://f:/c') is '["http:","f","","/c","",""]'
PASS segments('http://f:0/c') is '["http:","f","0","/c","",""]'
PASS segments('http://f:00000000000000/c') is '["http:","f","0","/c","",""]'
FAIL segments('http://f:00000000000000000000080/c') should be ["http:","f","0","/c","",""]. Was ["http:","f","","/c","",""].
PASS segments('http://f:b/c') is '[":","","","","",""]'
PASS segments('http://f: /c') is '[":","","","","",""]'
PASS segments('http://f:\n/c') is '["http:","f","","/c","",""]'
PASS segments('http://f:fifty-two/c') is '[":","","","","",""]'
FAIL segments('http://f:999999/c') should be [":","","0","","",""]. Was [":","","","","",""].
PASS segments('http://f: 21 / b ? d # e ') is '[":","","","","",""]'
FAIL segments('') should be ["data:","","","text/plain,baseURL","",""]. Was [":","","","","",""].
FAIL segments('  \t') should be ["data:","","","text/plain,baseURL","",""]. Was [":","","","","",""].
PASS segments(':foo.com/') is '[":","","","","",""]'
PASS segments(':foo.com\\') is '[":","","","","",""]'
PASS segments(':') is '[":","","","","",""]'
PASS segments(':a') is '[":","","","","",""]'
PASS segments(':/') is '[":","","","","",""]'
PASS segments(':\\') is '[":","","","","",""]'
PASS segments(':#') is '[":","","","","",""]'
FAIL segments('#') should be [":","","","","",""]. Was ["data:","","","text/plain,baseURL","",""].
FAIL segments('#/') should be [":","","","","",""]. Was ["data:","","","text/plain,baseURL","","#/"].
FAIL segments('#\\') should be [":","","","","",""]. Was ["data:","","","text/plain,baseURL","","#\\"].
FAIL segments('#;?') should be [":","","","","",""]. Was ["data:","","","text/plain,baseURL","","#;?"].
PASS segments('?') is '[":","","","","",""]'
PASS segments('/') is '[":","","","","",""]'
PASS segments(':23') is '[":","","","","",""]'
FAIL segments('/:23') should be ["data:","","","/:23","",""]. Was [":","","","","",""].
PASS segments('//') is '[":","","","","",""]'
PASS segments('::') is '[":","","","","",""]'
PASS segments('::23') is '[":","","","","",""]'
FAIL segments('foo://') should be ["foo:","","","//","",""]. Was ["foo:","","","","",""].
PASS segments('http://a:b@c:29/d') is '["http:","c","29","/d","",""]'
PASS segments('http::@c:29') is '["http:","c","29","/","",""]'
PASS segments('http://&a:foo(b]c@d:2/') is '["http:","d","2","/","",""]'
PASS segments('http://::@c@d:2') is '["http:","d","2","/","",""]'
PASS segments('http://foo.com:b@d/') is '["http:","d","","/","",""]'
PASS segments('http://foo.com/\\@') is '["http:","foo.com","","//@","",""]'
PASS segments('http:\\\\foo.com\\') is '["http:","foo.com","","/","",""]'
PASS segments('http:\\\\a\\b:c\\d@foo.com\\') is '["http:","a","","/b:c/d@foo.com/","",""]'
PASS segments('foo:/') is '["foo:","","","/","",""]'
PASS segments('foo:/bar.com/') is '["foo:","","","/bar.com/","",""]'
FAIL segments('foo://///////') should be ["foo:","","","/////////","",""]. Was ["foo:","","","///////","",""].
FAIL segments('foo://///////bar.com/') should be ["foo:","","","/////////bar.com/","",""]. Was ["foo:","","","///////bar.com/","",""].
FAIL segments('foo:////://///') should be ["foo:","","","////://///","",""]. Was ["foo:","","","//://///","",""].
PASS segments('c:/foo') is '["c:","","","/foo","",""]'
PASS segments('//foo/bar') is '[":","","","","",""]'
PASS segments('http://foo/path;a??e#f#g') is '["http:","foo","","/path;a","??e","#f#g"]'
PASS segments('http://foo/abcd?efgh?ijkl') is '["http:","foo","","/abcd","?efgh?ijkl",""]'
PASS segments('http://foo/abcd#foo?bar') is '["http:","foo","","/abcd","","#foo?bar"]'
FAIL segments('[61:24:74]:98') should be ["data:","","","text/[61:24:74]:98","",""]. Was [":","","","","",""].
FAIL segments('http://[61:27]:98') should be [":","","0","","",""]. Was [":","","","","",""].
PASS segments('http:[61:27]/:foo') is '[":","","","","",""]'
PASS segments('http://[1::2]:3:4') is '[":","","","","",""]'
PASS segments('http://2001::1') is '[":","","","","",""]'
PASS segments('http://[2001::1') is '[":","","","","",""]'
PASS segments('http://2001::1]') is '[":","","","","",""]'
PASS segments('http://2001::1]:80') is '[":","","","","",""]'
PASS segments('http://[2001::1]') is '["http:","[2001::1]","","/","",""]'
PASS segments('http://[2001::1]:80') is '["http:","[2001::1]","","/","",""]'
PASS segments('http://[[::]]') is '[":","","","","",""]'
PASS successfullyParsed is true

TEST COMPLETE
